extend elevator to support different types of elevators
